+++ backend/src/main/java/group8/bloodbank/model/Donor.java
@@ -18,7 +18,7 @@ public class Donor extends User {
     private Category category;
     private BloodType bloodType;
 
-    public Donor(int id, String name, String surname, String password, Address address, String jmbg, String email, String occupation, int penalty, Gender gender) {
+    public Donor(Long id, String name, String surname, String password, Address address, String jmbg, String email, String occupation, int penalty, Gender gender) {
         super(id, name, surname, password, address, jmbg, email, occupation, penalty, gender);
     }
