+++ backend/src/main/java/group8/bloodbank/model/MedicalWorker.java
@@ -14,7 +14,7 @@ public class MedicalWorker extends User {
      */
     public BloodBank bloodBank;
 
-    public MedicalWorker(int id, String name, String surname, String password, Address adress, String jmbg, String email, String occupation, int penalty, Gender gender) {
+    public MedicalWorker(Long id, String name, String surname, String password, Address adress, String jmbg, String email, String occupation, int penalty, Gender gender) {
         super(id, name, surname, password, adress, jmbg, email, occupation, penalty, gender);
 
     }
